Fix remaining Linux/Windows warnings
GML_Base.h / GML_Base.cpp:
  optimization_problem(double*) → (const double*) to match
  Decomposition_Base virtual, fixes -Woverloaded-virtual on Linux/GCC

Variational_Quantum_Eigensolver_Base.h:
  Same const double* fix for the sibling overload declaration

GML_Base.cpp:
  Wrap MSVC-only '#pragma warning(suppress: 4996)' in #ifdef _WIN32
  to fix -Wunknown-pragmas on GCC/Clang

density_matrix/CMakeLists.txt:
  Remove explicit /W4 from MSVC target_compile_options — the root
  CMakeLists.txt already sets /W3, and adding /W4 here caused the
  D9025 'overriding /W3 with /W4' command-line warning on Windows

density_matrix/bindings.cpp:
  static_cast<int>(buf.size) for NoisyCircuit::apply_to call,
  fixes C4244 pybind11::ssize_t→int on Windows

+++ squander/density_matrix/bindings.cpp
@@ -406,7 +406,7 @@ PYBIND11_MODULE(_density_matrix_cpp, m) {
           [](NoisyCircuit &self, py::array_t<double> params,
              DensityMatrix &rho) {
             auto buf = params.request();
-            self.apply_to(static_cast<double *>(buf.ptr), buf.size, rho);
+            self.apply_to(static_cast<double *>(buf.ptr), static_cast<int>(buf.size), rho);
           },
           py::arg("parameters"), py::arg("density_matrix"),
           "Apply circuit to density matrix")
